+++ frontend-ts/src/components/SelectedMoviesSection/index.tsx
@@ -3,17 +3,10 @@ import Paper from '@mui/material/Paper'
 import Stack from '@mui/material/Stack'
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
-import type { MovieCardSelectedProps } from '../MovieCardSelected'
+import type { SelectedMoviesSectionProps } from '../../types'
 import noMoviesImageSrc from '../../assets/no_movies.png'
 import { MovieCardSelected } from '../MovieCardSelected'
 import SelectedMoviesForm from '../SelectedMoviesForm'
-
-type MovieType = MovieCardSelectedProps['movie']
-
-export type SelectedMoviesSectionProps = {
-  selectedMovies: MovieType[]
-  deleteMovie: (selectedMovie: MovieType) => void
-}
 
 const SelectedMovies = styled(Paper)(({ theme }) => ({
   backgroundColor: '#fff',
