+++ frontend-ts/src/types/index.tsx
@@ -31,7 +31,7 @@ type Genres = {
   name: string
 }
 
-type MovieType2 = {
+type MovieSelectedType = {
   title?: string
   image?: string
   releaseDate?: string
@@ -41,11 +41,31 @@ type MovieType2 = {
 }
 
 export type MovieCardSelectedProps = {
-  movie: MovieType2
+  movie: MovieSelectedType
   display?: 'flex' | 'block'
   /** what the type of minheight is
    * @default 164
    */
   minHeight?: number
-  onCardDelete: (movie: MovieType2) => void
+  onCardDelete: (movie: MovieSelectedType) => void
+}
+
+/** Navigation component */
+export type NavigationProps = {
+  /** the default number
+   * @default 1
+   */
+  flexGrow: number
+  /** width number
+   * @default 250
+   */
+  width: number
+}
+
+/** SelectedMoviesSection component */
+type MovieSelectedType2 = MovieCardSelectedProps['movie']
+
+export type SelectedMoviesSectionProps = {
+  selectedMovies: MovieSelectedType2[]
+  deleteMovie: (selectedMovie: MovieType) => void
 }
